refactor: TopTabs 폴더 이동 및 코드 리뷰 수정
