+++ src/domain/Roulette/components/RouletteFilter/types.ts
@@ -1,0 +1,6 @@
+import { MenuItem } from '../../utils/menuItem';
+
+export interface RouletteFilterProps {
+  onChange: (menus: MenuItem[]) => void;
+  disabled?: boolean;
+}
