+++ src/domain/Roulette/components/RouletteUi/types.ts
@@ -1,4 +1,4 @@
-import { MenuItem } from './../../utils/menuItem';
+import { MenuItem } from '../../utils/menuItem';
 export interface RouletteUiProps {
   items: MenuItem[]; // 룰렛에 표시될 항목 목록
   onResult?: (item: MenuItem) => void; // 룰렛이 멈췄을 때 선택된 항목을 전달하는 콜백
