+++ src/domain/TopTabs/TabButton/TabButton.module.scss
