+++ src/domain/TopTabs/TopTabs.module.scss
